Extracted Contentproviders to WorkflowstateHandler
